repro-explain: move unit tests out of source files
