+++ Cargo.toml
@@ -32,8 +32,8 @@ members = [
   "src/tools/opt-dist",
   "src/tools/remote-test-client",
   "src/tools/remote-test-server",
-  "src/tools/repro-explain",
   "src/tools/replace-version-placeholder",
+  "src/tools/repro-explain",
   "src/tools/run-make-support",
   "src/tools/rust-installer",
   "src/tools/rustdoc",
